Make the skippable parser's bound check independent of size_t width
It compared in 64 bits to stop GCOMP_LZ4_SKIPPABLE_OVERHEAD + declared
from wrapping where size_t is 32 bits wide.  That guard is real, but this
machine has no 32-bit toolchain, so nothing in the suite could exercise
it -- reintroducing the wrap as a deliberate defect changed no test
result, which is the one falsification in this stretch that found
nothing.

Subtracting from the input size instead of adding to the declared size is
exact at either width.  input_size is already known to be at least
GCOMP_LZ4_SKIPPABLE_OVERHEAD, so the subtraction cannot underflow, and
the correctness no longer rests on a guard only one platform reaches.

The writer's payload-size check keeps its 64-bit comparison: there the
widening is what stops a 32-bit build from seeing a tautology and
failing under -Werror.

+++ src/methods/lz4/lz4_frame.c
@@ -270,11 +270,15 @@ gcomp_status_t gcomp_lz4_read_skippable_frame(const void * input,
 
   uint32_t declared = gcomp_read_le32(in + 4);
 
-  // Compare in 64 bits.  The declared size is attacker-controlled and
-  // GCOMP_LZ4_SKIPPABLE_OVERHEAD + declared would wrap on a 32-bit size_t,
-  // turning a frame that claims nearly 4 GB into one that appears to fit.
-  if ((uint64_t)declared >
-      (uint64_t)input_size - (uint64_t)GCOMP_LZ4_SKIPPABLE_OVERHEAD) {
+  // Subtract from the input size rather than adding to the declared one.
+  // The declared size is attacker-controlled and reaches 0xFFFFFFFF, so
+  // `GCOMP_LZ4_SKIPPABLE_OVERHEAD + declared` would wrap where size_t is 32
+  // bits wide and a frame claiming nearly 4 GB would appear to fit.  This
+  // subtraction cannot underflow -- input_size is already known to be at
+  // least GCOMP_LZ4_SKIPPABLE_OVERHEAD -- and is exact at either width, so
+  // the correctness does not rest on a guard that only one platform
+  // exercises.
+  if ((size_t)declared > input_size - GCOMP_LZ4_SKIPPABLE_OVERHEAD) {
     return GCOMP_ERR_CORRUPT;
   }
